requirements.txt
torch~=2.1.2
torchvision~=0.16.2
lightning~=2.1.3
pillow~=10.2.0
torchmetrics~=1.3.0.post0
tensorboard